Felhantering med trycatch och loggning med SLF4J

+++ src/main/java/Menu/RemoveKandidat.java
@@ -1,11 +1,14 @@
 package Menu;
 
 import Models.KandidatRepo;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.util.Scanner;
 
 public class RemoveKandidat implements Entry {
     private final KandidatRepo kandidatRepo;
+    Logger logger = LoggerFactory.getLogger(RemoveKandidat.class);
 
     public RemoveKandidat(KandidatRepo kandidatRepo) {
         this.kandidatRepo = kandidatRepo;
@@ -23,7 +26,11 @@ public class RemoveKandidat implements Entry {
         System.out.println("Skriv namn:");
         String name = scanner.nextLine();
 
-        kandidatRepo.removeKandidat(name);
-        System.out.println("Kandidat borttagen");
+       if(!kandidatRepo.kandidatList.containsKey(name)){
+           logger.info("Kandidat finns inte");
+       }else{
+           kandidatRepo.removeKandidat(name);
+           System.out.println("Kandidat borttagen");
+       }
     }
 }
